Unit test organizing:
- moved unit tests using any class dependent on BaseClasses to
'BaseClassDependentUnitTests'
- created common package under character_sheet_unit_tests
- moved method used by both CharacterSheetUnitTests and
BaseClassDependentUnitTests to CommonMethods under common package for
easier manageability
- added new unit test class to CharacterSheetTestSuite.java
