+++ RPGManager_development/test/unit_test/test_suits/CharacterSheetTestSuite.java
@@ -3,6 +3,7 @@ package unit_test.test_suits;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
+import unit_test.character_sheet_unit_tests.BaseClassDependentUnitTests;
 import unit_test.character_sheet_unit_tests.CharacterSheetUnitTests;
 import unit_test.character_sheet_unit_tests.InvalidBackgroundUnitTests;
 import unit_test.character_sheet_unit_tests.InvalidClassAndSpecializationPairingsTests;
@@ -10,8 +11,8 @@ import unit_test.character_sheet_unit_tests.ValidBackgroundUnitTests;
 import unit_test.character_sheet_unit_tests.ValidClassAndSpecializationPairingsTests;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({ CharacterSheetUnitTests.class, InvalidBackgroundUnitTests.class, InvalidClassAndSpecializationPairingsTests.class,
-		ValidBackgroundUnitTests.class, ValidClassAndSpecializationPairingsTests.class })
+@Suite.SuiteClasses({ CharacterSheetUnitTests.class, BaseClassDependentUnitTests.class, InvalidBackgroundUnitTests.class,
+		InvalidClassAndSpecializationPairingsTests.class, ValidBackgroundUnitTests.class, ValidClassAndSpecializationPairingsTests.class })
 
 public class CharacterSheetTestSuite {
